cmake: Warn that MSVC is unmaintained.
It has been broken for a while, and not much interest from our side
in keeping it going, as most of us are using MinGW cross-compilers
anyway.

I'm not sure if there are other people interested on MSVC, so just
warn for the time being.

Also tone down the MSVC warnings, so they don't overwhelm the error
messages.

Reviewed-by: Brian Paul <brianp@vmware.com>
Reviewed-by: Roland Scheidegger <sroland@vmware.com>