wglgears: mark application as dpi-aware
This makes rendering nice and crisp on high-dpi systems. Since we're not
doing any layouting or anything, we can just claim to be aware without
doing much.

We keep the current window-size when the DPI changes, because the size
is specified in pixels, so it seems more useful to keep that as-is.

Reviewed-by: Brian Paul <brianp@vmware.com>