wglgears: fix rotation speed to 70 degrees per second
This copies the logic from glxgears, omitting the rotating-state,
because we have no such option (yet).

In order to do this, we also need to swap out our timer. In order to
keep things nice and easy, we can use timeGetTime(), but set the minumum
period before the event-loop.

And for the timing-calculation, this is a lot better with proper
resolution anyway.

Reviewed-by: Brian Paul <brianp@vmware.com>