wglgears: set whole process as dpi-aware
It seems different GPU drivers behave differently in the case of
per-thread DPI-awareness, causing incorrect rendering on drivers like
the NVIDIA-driver. This is probably related to some internal threading
in the drivers, and is probably a subtle driver-bug.

In either case, setting per-process DPI awareness seems to do the trick.
It doesn't matter to us if we're doing per process or per thread, so
let's just go with the more robust solution.

Fixes: 7fa3a433 ("wglgears: mark application as dpi-aware")