wglgears: omit paranoid error-checking
This code contains a lot of code that can't realistically fail; these
things can only fail on incorrect input, or under memory pressure. We're
passing correct input, and we're not yet at a point where memory
pressure is really possible.

So for the sake of readability, let's remove some overly paranoid
error-checking. It's not like these error-messages are all that useful
to most users, compared to running this in a debugger anyway.

This makes the code reasier to read, and makes it easier to add
meaningful error-checking where it actually matters in the future.

Reviewed-by: Brian Paul <brianp@vmware.com>