wglgears: do not re-create window after re-creating context
When we re-created the window, we accidentally leaked the old window.
It's not so easy to notice, because it's invisible, but it's still there
and taking up some memory.

But there's more problems; when I recently added the -geometry switch,
I missed the extra CreateWindowEx call, meaning this call didn't respect
the x and y offset.

But instead of fixing these things indevidually, let's take a step back
here. There's no reason to re-create the window here in the first place.
We can just set a new pixel-format on hDC and re-create hRC instead.

This makes the code both easier to maintain, and squashes a couple of
bugs.

Reviewed-by: Brian Paul <brianp@vmware.com>