Revert "wglgears: do not re-create window after re-creating context"
This fix wasn't a great idea; even if it seems to work fine on modern
Windows versions, the documentation says that SetPixelFormat can only be
set once per window. And worse, Mesa's OpenGL32.DLL emulation code seems
to depend on this, ignoring any consecutive calls. This means that with
Mesa, we get no multisampling here.

So, let's revert this, and instead follow what the docs say.

This reverts commit 27813895adef5866e9aee62975c7f5965cd9c601.

Reviewed-by: Brian Paul <brianp@vmware.com>