egl: pack all strips per gear into a single draw
We can pack multiple triangle-strips into a single draw, by inserting
degenerate triangles between them. This should be a big win in terms of
CPU usage.

To avoid inserting needless degenerate triangles to invert face-winding,
we can manually invert the order of the back-face as well as emit the
inner face last. This ensures every strip starts with the right
face-winding.

Reviewed-by: Alexandros Frantzis <alexandros.frantzis@collbora.com>